Extract tests, cover multiple error path
