+++ lib/cookie.js
@@ -1418,7 +1418,7 @@ CookieJar.prototype.removeAllCookies = function(cb) {
       completedCount++;
 
       if (completedCount === cookies.length) {
-        return removeErrors.length > 0 ? cb(removeErrors[0]) : cb(null);
+        return cb(removeErrors.length ? removeErrors[0] : null);
       }
     }
